Первая итерация

+++ src/main/java/model/Apple.java
@@ -3,7 +3,7 @@ package model;
 public class Apple extends Food {
     String colour;
 
-    public Apple(int amount, double price, String colour, boolean isVegetarian, int discount) {
+    public Apple(int amount, double price, String colour, boolean isVegetarian, double discount) {
         this.amount = amount;
         this.price = price;
         this.colour = colour;
