+++ src/main/java/model/constants/Discount.java
@@ -1,5 +1,5 @@
 package model.constants;
 
 public class Discount {
-    public static final int DISCOUNT = 60;
+    public static final double DISCOUNT = 60;
 }
